Bonooru's Song Setup
An attempt at syncing the Scarecrow's Song for players who have yet to visit and do so themselves. Incomplete due to missing emits for notifications, a way to broadcast what notes were taught to Bonooru to players affected by the sync, and  a proper check for the relevant flag for the Adult half of the process. Meant to work on changes to the OoTO core in its own commits on its own branch.